requirements.txt
aiohttp==3.7.4.post0
APScheduler==3.6.3
async-timeout==3.0.1
attrs==19.3.0
cachetools==4.2.1
certifi==2019.11.28
cffi==1.14.5
chardet==4.0.0
Click==7.0
colorhash==1.0.2
configparser==4.0.2
cycler==0.10.0
decorator==4.4.2
dnspython==2.1.0
Flask==1.1.1
Flask-Cors==3.0.8
Flask-MonitoringDashboard==3.0.6
fsspec==0.8.7
gcsfs==0.7.2
google-api-core==1.26.1
google-auth==1.28.0
google-auth-oauthlib==0.4.3
google-cloud-core==1.6.0
google-cloud-storage==1.36.2
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
gunicorn==20.0.4
idna==2.10
idna-ssl==1.1.0
imbalanced-learn==0.6.1
imblearn==0.0
importlib-metadata==1.4.0
itsdangerous==1.1.0
Jinja2==2.11.0
joblib==0.14.1
jsonschema==3.2.0
kiwisolver==1.1.0
kneed==0.5.1
MarkupSafe==1.1.1
matplotlib==3.1.2
more-itertools==8.1.0
multidict==5.1.0
numpy==1.18.1
oauthlib==3.1.0
packaging==20.9
pandas==0.25.3
protobuf==3.15.6
psutil==5.6.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pymongo==3.11.3
pyparsing==2.4.6
pyrsistent==0.15.7
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
regexp==0.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.22.1
scipy==1.4.1
six==1.14.0
sklearn==0.0
sklearn-pandas==1.8.0
SQLAlchemy==1.3.13
typing-extensions==3.7.4.3
tzlocal==2.0.0
urllib3==1.26.4
Werkzeug==0.16.1
wincertstore==0.2
xgboost==0.90
yarl==1.6.3
zipp==2.0.1
